refactor: 统一类型定义并优化导入路径
将类型定义集中到 `type_hints.py` 文件中，减少代码重复并提高可维护性。同时，将 `BinaryReader` 的导入路径从 `tools` 模块移动到 `core` 模块，保持代码结构的一致性。
