+++ src/league_tools/utils/type_hints.py
@@ -4,9 +4,16 @@
 # @Site    : x-item.com
 # @Software: Pycharm
 # @Create  : 2024/3/12 12:39
-# @Update  : 2024/5/5 2:18
+# @Update  : 2025/4/26 3:26
 # @Detail  : 
 
-from typing import Union
+import os
+from io import BytesIO, IOBase
+from typing import Union, List, TypeVar
 
 StrPath = Union[str, 'os.PathLike[str]']
+
+# 定义类型变量简化类型标注
+T = TypeVar('T')
+DataSource = Union[str, bytes, BytesIO, IOBase, os.PathLike]
+BinaryData = Union[bytes, bytearray, List[int], str]
